+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename experiment code to study period ID. [#456](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/456)
 - Remove people tab from experiment plans. [#459](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/459)
 - Improve sorting of plots. [#387](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/387)
+- Move field description for multiple value fields to the table caption. [#455](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/455)
 
 ### Fixed