+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.14.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/32) 2023-07-17
+
 ### Added
 
 - Add planting and harvest year to proposals. [#481](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/481)